Added INC case

+++ CPU.h
@@ -55,7 +55,7 @@ class CPU {
     uint8_t get_memory(addressing_mode, uint8_t);
     uint8_t get_memory(addressing_mode, uint8_t, uint8_t);
 
-    uint8_t make_address(addressing_mode, uint8_t);
+    uint16_t make_address(addressing_mode, uint8_t);
     uint16_t make_address(addressing_mode, uint8_t, uint8_t);
 
     // Given an address, execute the opcode at that address
